Added basic config support,cleaned up, added some comments...
So many minor changes!
You can set the origin of a sprite now!

+++ Engine/Text.cpp
@@ -15,37 +15,37 @@ void Text::operator=(std::string rhs) {
 void Text::setText(std::string str) {
 	_len = str.length();
 	indices = new char[_len]; 
-	for (int i = 0; i < _len; i++)
+	for (unsigned int i = 0; i < _len; i++)
 		indices[i] = str[i] - 32;
 
 	Sprite *l = new Sprite[_len];
-	for (int i = 0; i < _len; i++) {
+	for (unsigned int i = 0; i < _len; i++) {
 		if (indices[i] > -1) {
 			Sprite character;
-			character.init(x,y,size,size,false,texture);
+			character.init(x,y,(float)size,(float)size,false,texture);
 			character.swapUVs(indices[i]);
-			character.setPosition(i*size, 64);
+			character.setPosition((float)i*size, 64);
 		}
 	}
 }
 
 void Text::init(int pos_x, int pos_y, char tex[],int fontSize) {
 	size = fontSize;
-	x = pos_x;
-	y = pos_y;
+	x = (float)pos_x;
+	y = (float)pos_y;
 	texture = tex;
 
 	font.UVGridDivisions = 16;
-	font.init(x, y, size, size, false, texture);
+	font.init(x, y, (float)size, (float)size, false, texture);
 }
 
 void Text::render(GLSLShading shader, glm::mat4 cMat) {
 	glActiveTexture(GL_TEXTURE0);
 	glBindTexture(GL_TEXTURE_2D, font.getTextureID());
-	for (int i = 0; i < _len; i++) {
+	for (unsigned int i = 0; i < _len; i++) {
 		if (indices[i] > -1) {
 			font.swapUVs(indices[i]);
-			font.setPosition(i*size,64);
+			font.setPosition((float)i*size,64);
 			font.render(shader, cMat);
 		}
 	}
